% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentClassJM.R
\name{create.data}
\alias{create.data}
\title{Generate Dataset}
\usage{
create.data(n, seed = 1)
}
\arguments{
\item{n}{Sample size}

\item{seed}{Seed of the random generator (optional)}
}
\value{
A list of the following components: \itemize{  
  \item \strong{Y} :  An \eqn{(n \times J \times m)} array of longitudinal outcome measurements, where \eqn{n} is the sample size, \eqn{J} is the number of longitudinal measurement types, and
\eqn{m} is the maximum number of measurement times. It can contain \code{NA} values if the number of measurements for a subject is fewer than the maximum number of measurements. The \eqn{(i,j,k)}th element corresponds to the \eqn{k}th measurement of the \eqn{j}th type of longitudinal outcome for the \eqn{i}th subject
  \item \strong{X} : An \eqn{(n \times J \times m \times p_X)} array of covariates (excluding intercept) of the longitudinal outcome model, 
where \eqn{n} is the sample size, \eqn{J} is the number of longitudinal measurement types, \eqn{m} is the number of measurement times, and \eqn{p_X} is the number of covariates. 
The \eqn{(i,j,k,l)}th element corresponds to the \eqn{l}th covariate for the \eqn{k}th measurement of the \eqn{j}th type of longitudinal outcome for the \eqn{i}th subject
  \item \strong{W} : An \eqn{(n \times p_W)} matrix of covariates for the latent class regression model, where \eqn{n} is the sample size, and 
\eqn{p_W} is the number of covariate. The \eqn{(i,l)}th element corresponds to the \eqn{l}th covariate for the \eqn{i}th subject
  \item \strong{Time} : An \eqn{n}-vector of observed event or censoring times
  \item \strong{D} : An \eqn{n}-vector of event indicators
  \item \strong{ni} : An \eqn{(n \times J)} matrix of numbers of measurements for the longitudinal outcomes
  \item \strong{Z} : An \eqn{(n \times p_Z)} matrix of time-independent covariates for the survival model, where \eqn{n} is the sample size, and 
\eqn{p_Z} is the number of covariates. The \eqn{(i,l)}th element corresponds to the \eqn{l}th covariate for the \eqn{i}th subject
  }
  Based on the setting of the simulation studies in Wong et al. (2022), we fix \eqn{J = 2}, \eqn{m = 10}, \eqn{p_X = 3}, \eqn{p_W = 2}, and \eqn{p_Z = 2}.
}
\description{
Create a dataset based on the setting of the simulation studies in Wong et al. (2022)
}
\examples{
dataset <- create.data(n=1000)
}
\references{
Wong, K. Y., Zeng, D., & Lin, D. Y. (2022). Semiparametric latent-class models for multivariate longitudinal and survival data. The Annals of Statistics. 50 487–510.
}
